+++ src/Client/ClientSetup.h
@@ -1,3 +1,5 @@
+#pragma once
+
 #include "Client.h"
 
 class ClientUtils